Run all providers on start
